// src/common/enums/brazilian-state.enum.ts
/**
 * Estados do Brasil (nomes completos). Use estes valores em endereços (state).
 * Não use abreviações (UF).
 */
export enum BrazilianState {
  Acre = 'Acre',
  Alagoas = 'Alagoas',
  Amapa = 'Amapá',
  Amazonas = 'Amazonas',
  Bahia = 'Bahia',
  Ceara = 'Ceará',
  DistritoFederal = 'Distrito Federal',
  EspiritoSanto = 'Espírito Santo',
  Goias = 'Goiás',
  Maranhao = 'Maranhão',
  MatoGrosso = 'Mato Grosso',
  MatoGrossoDoSul = 'Mato Grosso do Sul',
  MinasGerais = 'Minas Gerais',
  Para = 'Pará',
  Paraiba = 'Paraíba',
  Parana = 'Paraná',
  Pernambuco = 'Pernambuco',
  Piaui = 'Piauí',
  RioDeJaneiro = 'Rio de Janeiro',
  RioGrandeDoNorte = 'Rio Grande do Norte',
  RioGrandeDoSul = 'Rio Grande do Sul',
  Rondonia = 'Rondônia',
  Roraima = 'Roraima',
  SantaCatarina = 'Santa Catarina',
  SaoPaulo = 'São Paulo',
  Sergipe = 'Sergipe',
  Tocantins = 'Tocantins',
}